Add constants

+++ Sources/MoexDto/WS/Properties.swift
@@ -1,6 +1,6 @@
 //
 //  Properties.swift
-//  
+//  MoexDto
 //
 //  Created by Victor Chernykh on 22.07.2024.
 //
@@ -8,7 +8,7 @@
 /// Model Properties.
 public struct Properties: Codable, Sendable {
 	public let type: PropertyType
-	public let seqnum: Int
-	public let published: Int?
-	public let timestamp: Int
+	public let seqnum: UInt64
+	public let published: UInt64?
+	public let timestamp: UInt64
 }
